LV: Types and code cleanup. Add version switcher to more places.
